Refactored withValidation to fix defaultProps problem with validated components
